update 4.1.8
Added a "Music" tab. From there, simply by selecting a folder that had .mp3 or .m4a music files, you can play them from there. It also has a pause, play, next, previous song functions and when the music is playing, after the song finishes it goes to the next song. If it reaches the end of the song list it starts over again.