some changes to LaunchGame() and SocialMediaLoader()
in this new version, the "plyer" library will not be a requirement anymore because it is not used. For some reason, plyer notifications work while the app is a .py, .pyw but now when its compiled into a .exe. Couldn't really figure out why so i changed the functions a bit. I also changed how the buttons get constructed in the AppsLaucherGUISetup() function. Still same concept but now it also passes the "url_name" param to the functions.