checkpoint
An improvement in the update_all_music_frame() function. Instead of destroying the widget then creating it again with the new data, now it grid_forget()'s the widgets making them disappear from the frame then adding that widget to the delete_widgets list. Then it goes and creates the new widgets with the new data and finally it deletes the old widgets in the delete_widgets list. The errors that i was encountering were about drawing to the screen/frame so i thought that maybe it wasn't able to display the removal and the new items quick enough therefore throwing an exception. Going to look into this more and see if i can perfect the functionality. Before, it took about 10-15 seconds to scan a local directory and actually get ready to play the songs in your selected directory. Now it takes 0.7 seconds to scan the directory and about 2-3 seconds to get everything on the GUI to show properly (i am testing the performance with 308 songs in a directory (1.69GB). I have tested both on a local directory and a SMB share. Each of the songs length vary from 1 minute to 8 minutes)