checkpoint
MusicManager is pretty much ready to be used but i want to stress test it before i release anything. The are a lot of differences between the old version and this version but the first thing that you'll probably realize is it loads songs a lot more faster then before. It also has support to load any file types that are in TinyTag.SUPPORTED_FILE_EXTENSIONS. Instead of using pygame to manage the music (play, pause, next, etc) we now have vlc (python-vlc) doing the managing and TinyTag (tinytag) doing all the file searching. Still got a lot of work to do to get this 100% stable so i'll leave it at that.