checkpoint
You can now "stop" the song without closing out of the app. Before when you paused the currently playing song, VLC didn't release the file so you couldn't modify both the file itself and the directory it was in. So now we got a new button that actually releases the song.